+++ agent_llamaindex/custom_model/agent.py
@@ -67,6 +67,7 @@ class MyAgent:
         api_base: Optional[str] = None,
         model: Optional[str] = None,
         verbose: Optional[Union[bool, str]] = True,
+        timeout: Optional[int] = 90,
         **kwargs: Any,
     ):
         """Initializes the MyAgent class with API key, base URL, model, and verbosity settings.
@@ -80,6 +81,8 @@ class MyAgent:
                 Defaults to None.
             verbose: Optional[Union[bool, str]]: Whether to enable verbose logging.
                 Accepts boolean or string values ("true"/"false"). Defaults to True.
+            timeout: Optional[int]: How long to wait for the agent to respond.
+                Defaults to 90 seconds.
             **kwargs: Any: Additional keyword arguments passed to the agent.
                 Contains any parameters received in the CompletionCreateParams.
 
@@ -89,6 +92,7 @@ class MyAgent:
         self.api_key = api_key or os.environ.get("DATAROBOT_API_TOKEN")
         self.api_base = api_base or os.environ.get("DATAROBOT_ENDPOINT")
         self.model = model
+        self.timeout = timeout
         if isinstance(verbose, str):
             self.verbose = verbose.lower() == "true"
         elif isinstance(verbose, bool):
@@ -123,6 +127,7 @@ class MyAgent:
             model="datarobot/azure/gpt-4o-mini",
             api_base=self.api_base_litellm,
             api_key=self.api_key,
+            timeout=self.timeout,
         )
 
     @property
@@ -134,17 +139,27 @@ class MyAgent:
         and another for a specific DataRobot deployment, or even multiple deployments or
         third-party LLMs.
         """
+        deployment_url = f"{self.api_base}/deployments/{os.environ.get('LLM_DATAROBOT_DEPLOYMENT_ID')}/"
 
         # NOTE: LlamaIndex tool encodings are sensitive the the LLM model used and may need to be re-written
         # to work with different models. This example assumes the model is a GPT compatible model.
         return DataRobotLiteLLM(
-            model="datarobot/azure/gpt-4o-mini",
-            api_base=f"{self.api_base_litellm}/api/v2/deployments/{os.environ.get('LLM_DEPLOYMENT_ID')}/",
+            model="openai/gpt-4o-mini",
+            api_base=deployment_url,
             api_key=self.api_key,
+            timeout=self.timeout,
         )
 
+    @property
+    def llm(self) -> DataRobotLiteLLM:
+        """Returns a LlamaIndex LiteLLM compatible LLM instance configured to use DataRobot's LLM Gateway or a specific deployment."""
+        if os.environ.get("LLM_DATAROBOT_DEPLOYMENT_ID"):
+            return self.llm_with_datarobot_deployment
+        else:
+            return self.llm_with_datarobot_llm_gateway
+
     @staticmethod
-    async def record_notes(ctx: Context, notes: str, notes_title: str) -> str:  # type: ignore[type-arg]
+    async def record_notes(ctx: Context, notes: str, notes_title: str) -> str:
         """Useful for recording notes on a given topic. Your input should be notes with a
         title to save the notes under."""
         current_state = await ctx.get("state")
@@ -155,7 +170,7 @@ class MyAgent:
         return "Notes recorded."
 
     @staticmethod
-    async def write_report(ctx: Context, report_content: str) -> str:  # type: ignore[type-arg]
+    async def write_report(ctx: Context, report_content: str) -> str:
         """Useful for writing a report on a given topic. Your input should be a markdown formatted report."""
         current_state = await ctx.get("state")
         current_state["report_content"] = report_content
@@ -163,7 +178,7 @@ class MyAgent:
         return "Report written."
 
     @staticmethod
-    async def review_report(ctx: Context, review: str) -> str:  # type: ignore[type-arg]
+    async def review_report(ctx: Context, review: str) -> str:
         """Useful for reviewing a report and providing feedback. Your input should be a review of the report."""
         current_state = await ctx.get("state")
         current_state["review"] = review
@@ -181,7 +196,7 @@ class MyAgent:
                 "WriteAgent to write a report on the topic. You should have at least some notes on a topic "
                 "before handing off control to the WriteAgent."
             ),
-            llm=self.llm_with_datarobot_llm_gateway,
+            llm=self.llm,
             tools=[self.record_notes],
             can_handoff_to=["WriteAgent"],
         )
@@ -196,7 +211,7 @@ class MyAgent:
                 "Your report should be in a markdown format. The content should be grounded in the research notes. "
                 "Once the report is written, you should get feedback at least once from the ReviewAgent."
             ),
-            llm=self.llm_with_datarobot_llm_gateway,
+            llm=self.llm,
             tools=[self.write_report],
             can_handoff_to=["ReviewAgent", "ResearchAgent"],
         )
@@ -212,7 +227,7 @@ class MyAgent:
                 "WriteAgent to implement.  If you have feedback that requires changes, you should hand "
                 "off control to the WriteAgent to implement the changes after submitting the review."
             ),
-            llm=self.llm_with_datarobot_llm_gateway,
+            llm=self.llm,
             tools=[self.review_report],
             can_handoff_to=["WriteAgent"],
         )
@@ -268,7 +283,7 @@ class MyAgent:
 
     def run(
         self, completion_create_params: CompletionCreateParams
-    ) -> Tuple[str, Sequence[Event], dict[str, int]]:
+    ) -> Tuple[str, dict[str, int], Sequence[Event] | None]:
         """Run the agent with the provided completion parameters.
 
         [THIS METHOD IS REQUIRED FOR THE AGENT TO WORK WITH DRUM SERVER]
@@ -282,7 +297,8 @@ class MyAgent:
             completion_create_params (CompletionCreateParams): The parameters for
                 the completion request, which includes the input topic and other settings.
         Returns:
-            tuple[list[Any], CrewOutput]: A tuple containing a list of messages (events) and the crew output.
+            tuple[str, dict[str, int], Sequence[Event] | None]: A tuple containing the agent
+                response, usage_statistics and a list of messages (events).
 
         """
         # Example helper for extracting inputs as a json from the completion_create_params["messages"]
@@ -308,4 +324,9 @@ class MyAgent:
             "prompt_tokens": 0,
             "total_tokens": 0,
         }
-        return str(result["report_content"]), events, usage_metrics
+        # The `events` variable is used to compute agentic metrics
+        # (e.g. Task Adherence, Agent Goal Accuracy, Agent Goal Accuracy with Reference,
+        # Tool Call Accuracy).
+        # If you are not interested in these metrics, you can also return None instead.
+        # This will reduce the size of the response significantly.
+        return str(result["report_content"]), usage_metrics, events